requirements.txt
# langchain_community
langchain-openai
#langchainhub
langchain
langgraph
python-dotenv
IPython

